backend/requirements.txt
# FastAPI and ASGI server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Core dependencies
langchain==0.1.20
langchain-core==0.1.52
langchain-community==0.0.38
langchain-openai==0.1.7
langchain-pinecone==0.0.3
langsmith==0.1.31
python-dotenv


# Vector stores
pinecone-client>=3.0.0
docarray>=0.39.0

# Video processing
yt-dlp>=2023.0.0
openai-whisper>=20231117

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6

# Optional: For development
pytest>=7.4.0
httpx>=0.26.0  # For testing
black>=23.0.0
flake8>=6.0.0